+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/CallCompositeOptions/LocalizationOptions.swift
@@ -7,6 +7,10 @@ import SwiftUI
 
 /// SupportedLocale representing the supported locales.
 public struct SupportedLocale {
+    /// Arabic
+    public static let ar = Locale(identifier: "ar")
+    /// Arabic (Saudi Arabia)
+    public static let arSA = Locale(identifier: "ar-SA")
     /// Chinese, Simplified
     public static let zh = Locale(identifier: "zh")
     /// Chinese, Simplified
@@ -27,6 +31,10 @@ public struct SupportedLocale {
     public static let enGB = Locale(identifier: "en-GB")
     /// English (United States)
     public static let enUS = Locale(identifier: "en-US")
+    /// Finnish
+    public static let fi = Locale(identifier: "fi")
+    /// Finnish (Finland)
+    public static let fiFI = Locale(identifier: "fi-FI")
     /// French
     public static let fr = Locale(identifier: "fr")
     /// French (France)
@@ -35,6 +43,10 @@ public struct SupportedLocale {
     public static let de = Locale(identifier: "de")
     /// German (Germany)
     public static let deDE = Locale(identifier: "de-DE")
+    /// Hebrew
+    public static let he = Locale(identifier: "he")
+    /// Hebrew (Israel)
+    public static let heIL = Locale(identifier: "he-IL")
     /// Italian
     public static let it = Locale(identifier: "it")
     /// Italian (Italy)
@@ -47,6 +59,14 @@ public struct SupportedLocale {
     public static let ko = Locale(identifier: "ko")
     /// Korean (South Korea)
     public static let koKR = Locale(identifier: "ko-KR")
+    /// Norwegian Bokmål
+    public static let nb = Locale(identifier: "nb")
+    /// Norwegian Bokmål (Norway)
+    public static let nbNO = Locale(identifier: "nb-NO")
+    /// Polish
+    public static let pl = Locale(identifier: "pl")
+    /// Polish (Poland)
+    public static let plPL = Locale(identifier: "pl-PL")
     /// Portuguese
     public static let pt = Locale(identifier: "pt")
     /// Portuguese (Brazil)
@@ -59,6 +79,10 @@ public struct SupportedLocale {
     public static let es = Locale(identifier: "es")
     /// Spanish (Spain)
     public static let esES = Locale(identifier: "es-ES")
+    /// Swedish
+    public static let sv = Locale(identifier: "sv")
+    /// Swedish (Sweden)
+    public static let svSE = Locale(identifier: "sv-SE")
     /// Turkish
     public static let tr = Locale(identifier: "tr")
     /// Turkish (Turkey)