+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/DI/DependencyContainer.swift
@@ -52,13 +52,13 @@ final class DependencyContainer {
         register(AvatarViewManager(store: resolve(),
                                    localParticipantViewData: localOptions?.participantViewData) as
                  AvatarViewManagerProtocol)
-        register(DiagnosticsManager(store: resolve()) as DiagnosticsManagerProtocol)
+        register(DebugInfoManager(store: resolve()) as DebugInfoManagerProtocol)
         register(CompositeViewModelFactory(logger: resolve(),
                                            store: resolve(),
                                            networkManager: resolve(),
                                            localizationProvider: resolve(),
                                            accessibilityProvider: resolve(),
-                                           diagnosticsManager: resolve(),
+                                           debugInfoManager: resolve(),
                                            localOptions: localOptions) as CompositeViewModelFactoryProtocol)
         register(CompositeViewFactory(logger: resolve(),
                                       avatarManager: resolve(),
@@ -73,7 +73,6 @@ final class DependencyContainer {
                                      logger: resolve()) as AudioSessionManagerProtocol)
         register(RemoteParticipantsManager(store: resolve(),
                                            callCompositeEventsHandler: callCompositeEventsHandler,
-                                           callingSDKWrapper: resolve(),
                                            avatarViewManager: resolve()) as RemoteParticipantsManagerProtocol)
     }