+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/Factories/CompositeViewModelFactory.swift
@@ -17,11 +17,11 @@ protocol CompositeViewModelFactoryProtocol {
                                  isDisabled: Bool,
                                  action: @escaping (() -> Void)) -> IconButtonViewModel
     func makeIconWithLabelButtonViewModel<ButtonStateType>(
-                                 selectedButtonState: ButtonStateType,
-                                 localizationProvider: LocalizationProviderProtocol,
-                                 buttonTypeColor: IconWithLabelButtonViewModel<ButtonStateType>.ButtonTypeColor,
-                                 isDisabled: Bool,
-                                 action: @escaping (() -> Void)) -> IconWithLabelButtonViewModel<ButtonStateType>
+        selectedButtonState: ButtonStateType,
+        localizationProvider: LocalizationProviderProtocol,
+        buttonTypeColor: IconWithLabelButtonViewModel<ButtonStateType>.ButtonTypeColor,
+        isDisabled: Bool,
+        action: @escaping (() -> Void)) -> IconWithLabelButtonViewModel<ButtonStateType>
     func makeLocalVideoViewModel(dispatchAction: @escaping ActionDispatch) -> LocalVideoViewModel
     func makePrimaryButtonViewModel(buttonStyle: FluentUI.ButtonStyle,
                                     buttonLabel: String,
@@ -48,6 +48,7 @@ protocol CompositeViewModelFactoryProtocol {
     func makeLocalParticipantsListCellViewModel(localUserState: LocalUserState) -> ParticipantsListCellViewModel
     func makeParticipantsListCellViewModel(participantInfoModel: ParticipantInfoModel) -> ParticipantsListCellViewModel
     func makeMoreCallOptionsListViewModel(showSharingViewAction: @escaping () -> Void) -> MoreCallOptionsListViewModel
+    func makeDebugInfoSharingActivityViewModel() -> DebugInfoSharingActivityViewModel
     func makeDrawerListItemViewModel(icon: CompositeIcon,
                                      title: String,
                                      accessibilityIdentifier: String,
@@ -57,7 +58,6 @@ protocol CompositeViewModelFactoryProtocol {
         title: String,
         isSelected: Bool,
         onSelectedAction: @escaping (() -> Void)) -> SelectableDrawerListItemViewModel
-    func makeDiagnosticsSharingActivityViewModel() -> DiagnosticsSharingActivityViewModel
 
     // MARK: SetupViewModels
     func makePreviewAreaViewModel(dispatchAction: @escaping ActionDispatch) -> PreviewAreaViewModel
@@ -72,7 +72,7 @@ class CompositeViewModelFactory: CompositeViewModelFactoryProtocol {
     private let networkManager: NetworkManager
     private let accessibilityProvider: AccessibilityProviderProtocol
     private let localizationProvider: LocalizationProviderProtocol
-    private let diagnosticsManager: DiagnosticsManagerProtocol
+    private let debugInfoManager: DebugInfoManagerProtocol
     private let localOptions: LocalOptions?
 
     private weak var setupViewModel: SetupViewModel?
@@ -83,14 +83,14 @@ class CompositeViewModelFactory: CompositeViewModelFactoryProtocol {
          networkManager: NetworkManager,
          localizationProvider: LocalizationProviderProtocol,
          accessibilityProvider: AccessibilityProviderProtocol,
-         diagnosticsManager: DiagnosticsManagerProtocol,
+         debugInfoManager: DebugInfoManagerProtocol,
          localOptions: LocalOptions? = nil) {
         self.logger = logger
         self.store = store
         self.networkManager = networkManager
         self.accessibilityProvider = accessibilityProvider
         self.localizationProvider = localizationProvider
-        self.diagnosticsManager = diagnosticsManager
+        self.debugInfoManager = debugInfoManager
         self.localOptions = localOptions
     }
 
@@ -136,18 +136,18 @@ class CompositeViewModelFactory: CompositeViewModelFactoryProtocol {
                             action: action)
     }
     func makeIconWithLabelButtonViewModel<T: ButtonState>(
-                                          selectedButtonState: T,
-                                          localizationProvider: LocalizationProviderProtocol,
-                                          buttonTypeColor: IconWithLabelButtonViewModel<T>.ButtonTypeColor,
-                                          isDisabled: Bool,
-                                          action: @escaping (() -> Void)) -> IconWithLabelButtonViewModel<T> {
-        IconWithLabelButtonViewModel(
-                                     selectedButtonState: selectedButtonState,
-                                     localizationProvider: localizationProvider,
-                                     buttonTypeColor: buttonTypeColor,
-                                     isDisabled: isDisabled,
-                                     action: action)
-    }
+        selectedButtonState: T,
+        localizationProvider: LocalizationProviderProtocol,
+        buttonTypeColor: IconWithLabelButtonViewModel<T>.ButtonTypeColor,
+        isDisabled: Bool,
+        action: @escaping (() -> Void)) -> IconWithLabelButtonViewModel<T> {
+            IconWithLabelButtonViewModel(
+                selectedButtonState: selectedButtonState,
+                localizationProvider: localizationProvider,
+                buttonTypeColor: buttonTypeColor,
+                isDisabled: isDisabled,
+                action: action)
+        }
     func makeLocalVideoViewModel(dispatchAction: @escaping ActionDispatch) -> LocalVideoViewModel {
         LocalVideoViewModel(compositeViewModelFactory: self,
                             logger: logger,
@@ -273,9 +273,9 @@ class CompositeViewModelFactory: CompositeViewModelFactoryProtocol {
                                 action: action)
     }
 
-    func makeDiagnosticsSharingActivityViewModel() -> DiagnosticsSharingActivityViewModel {
-        DiagnosticsSharingActivityViewModel(accessibilityProvider: accessibilityProvider,
-                                            diagnosticsManager: diagnosticsManager)
+    func makeDebugInfoSharingActivityViewModel() -> DebugInfoSharingActivityViewModel {
+        DebugInfoSharingActivityViewModel(accessibilityProvider: accessibilityProvider,
+                                          debugInfoManager: debugInfoManager)
     }
 
     // MARK: SetupViewModels